% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datefixR-package.R
\docType{package}
\name{datefixR-package}
\alias{datefixR}
\alias{datefixR-package}
\title{datefixR: Fix Really Messy Dates}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

Fixes messy dates in data frames such as those entered via text boxes. Standardizes / - and whitespace separation, month abbreviations, and year first or day first by converting to R's built-in Date class. Imputes missing date or month using user-provided values.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.constantine-cooke.com/datefixR/ https://github.com/nathansam/datefixR}
  \item Report bugs at \url{https://github.com/nathansam/datefixR/issues}
}

}
\author{
\strong{Maintainer}: Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk} (\href{https://orcid.org/0000-0002-4437-8713}{ORCID})

}
\keyword{internal}
