% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datefixR-package.R
\docType{package}
\name{datefixR}
\alias{datefixR-package}
\alias{datefixR}
\title{datefixR: Standardize Dates in Different Formats or with Missing Data}
\description{
There are many different formats dates are commonly represented
with: the order of day, month, or year can differ, different separators
("-", "/", or whitespace) can be used, months can be numerical, names, or
abbreviations and year given as two digits or four. \code{datefixR} takes dates
in all these different formats and converts them to \R{}'s built-in date
class. If \code{datefixR} cannot standardize a date, such as because it is too
malformed, then the user is told which date cannot be standardized and the
corresponding ID for the row. \code{datefixR} also allows the imputation of
missing days and months with user-controlled behavior.

Get started by reading \code{vignette("datefixR")}
}
\seealso{
Useful links:
\itemize{
\item \url{https://docs.ropensci.org/datefixR/}
\item \url{https://github.com/ropensci/datefixR/}
\item Report bugs at \url{https://github.com/ropensci/datefixR/issues}
}
}
\author{
\strong{Maintainer}: Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk} (\href{https://orcid.org/0000-0002-4437-8713}{ORCID})

Authors:
\itemize{
  \item Nathan Constantine-Cooke \email{nathan.constantine-cooke@ed.ac.uk} (\href{https://orcid.org/0000-0002-4437-8713}{ORCID})
}

Other contributors:
\itemize{
  \item Jonathan Kitt \email{jonathan.kitt@protonmail.com} [contributor, translator]
  \item Antonio J. Pérez-Luque \email{ajpelu@gmail.com} (\href{https://orcid.org/0000-0002-1747-0469}{ORCID}) [contributor, translator]
  \item Daniel Possenriede \email{possenriede+r@gmail.com} (\href{https://orcid.org/0000-0002-6738-9845}{ORCID}) [contributor, translator]
  \item Michal Lauer \email{michal.lauer.25@gmail.com} [contributor, translator]
  \item Kaique dos S. Alves \email{kaiquedsalves@gmail.com} (\href{https://orcid.org/0000-0001-9187-0252}{ORCID}) [reviewer]
  \item Al-Ahmadgaid B. Asaad \email{alahmadgaid@gmail.com} (\href{https://orcid.org/0000-0003-3784-8593}{ORCID}) [reviewer]
  \item Anatoly Tsyplenkov \email{atsyplenkov@gmail.com} (\href{https://orcid.org/0000-0003-4144-8402}{ORCID}) [contributor, translator]
  \item Chitra M. Saraswati \email{chitra.m.saraswati@gmail.com} (\href{https://orcid.org/0000-0002-8159-0414}{ORCID}) [contributor, translator]
}

}
\keyword{internal}
