% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{exampledates}
\alias{exampledates}
\title{Example dataset of dates in different formats}
\format{
A data frame with 5 rows and 3 variables:
\describe{
\item{id}{Row ID (numeric).}
\item{some.dates}{Dates in different formats (character).}
\item{some.more.dates}{Additional dates in different formats (character).}
}
}
\usage{
exampledates
}
\description{
A toy dataset to use with datefixR functions.
}
\keyword{datasets}
