% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R, R/fix_date.R
\name{fix_date}
\alias{fix_date}
\title{Main date fixing function - Rust implementation of .fix_date}
\usage{
fix_date(date, day.impute = 1, month.impute = 7, format = "dmy")

fix_date(date, day.impute = 1, month.impute = 7, format = "dmy")
}
\arguments{
\item{date}{Character to be converted to \R{}'s \code{Date} class.}

\item{day.impute}{Integer between 1 and 31, or NA, or NULL. Day of the month
to be imputed when missing. Defaults to 1. If \code{day.impute = NA}, then
\code{NA} will be imputed for the date and a warning will be raised. If
\code{day.impute = NULL}, the function will fail with an error when day is
missing.}

\item{month.impute}{Integer between 1 and 12, or NA, or NULL. Month to be
imputed when missing. Defaults to 7 (July). If \code{month.impute = NA},
then \code{NA} will be imputed for the entire date and a warning will be
raised.
If \code{month.impute = NULL}, the function will fail with an error when
month is missing.}

\item{format}{Character string specifying date interpretation preference.
Either \code{"dmy"} (day-month-year, default) or \code{"mdy"}
(month-day-year, US format). This setting only affects ambiguous numeric
dates like "01/02/2023". When month names are present or year appears
first, the format is auto-detected regardless of this parameter. Note that
unambiguous dates (e.g., "25/12/2023") are parsed correctly regardless of
the format setting.}
}
\value{
An object belonging to \R's built-in \code{Date} class.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Converts a single improperly formatted date to \R{}'s Date class.
Supports numerous separators including /,- or white space.
Supports all-numeric, abbreviation or long-hand month notation. Where
day of the month has not been supplied, the first day of the month is
imputed. Either DMY or YMD is assumed by default. However, the US system of
MDY is supported via the \code{format} argument.
}
\examples{
bad.date <- "02 03 2021"
fixed.date <- fix_date(bad.date)
fixed.date
# - Due to deprecation, prefer using fix_date_char.
fixed.date <- fix_date_char(bad.date)
}
\seealso{
\code{\link{fix_date_char}} for character vectors and \code{\link{fix_date_df}} for data frames.
}
\keyword{internal}
