% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyapp.R
\name{fix_date_app}
\alias{fix_date_app}
\title{Shiny application standardizing date data in csv of excel files}
\usage{
fix_date_app(theme = "datefixR")
}
\arguments{
\item{theme}{Color theme for shiny app. Either \code{"datefixR"} (datefixR
colors) or \code{"none"}(default shiny app styling).}
}
\description{
Shiny app which allows users to standardize dates using
a graphical user interface (GUI). Most features of datefixR are supported
including imputed missing date data. Data can be provided as csv or xlsx
files. Processed datasets can be downloaded as csv files.
}
\examples{
\dontrun{
fix_date_app()
}
}
